Move group_by_nothing into odc.index
