+++ libs/index/odc/index/__init__.py
@@ -31,6 +31,11 @@ from ._yaml import (
     render_eo3_yaml,
 )
 
+from ._grouper import (
+    group_by_nothing,
+    solar_offset,
+)
+
 __all__ = (
     "from_yaml_doc_stream",
     "from_metadata_stream",
@@ -50,4 +55,6 @@ __all__ = (
     "utm_tile_dss",
     "mk_utm_gs",
     "render_eo3_yaml",
+    "group_by_nothing",
+    "solar_offset",
 )
